requirements/base.txt
Django==3.2.6
python-dotenv==0.15.0
mysqlclient==2.0.2
pytz==2020.5
django-allauth==0.45.0
cffi==1.14.6
cryptography==3.4.7
defusexml==0.7.1
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.1.0
python3-openid=3.2.0
requests-oauthlib==1.3.0
python-http-client==3.3.3
sendgrid==6.8.3
starkbank-ecdsa==1.1.1
django-sendgrid-v5==1.1.1
djangorestframework==3.12.4
django-cors-headers==3.10.0
